w3c-reports/12092014/gtk/runned-tests/websockets/interfaces/WebSocket/url/004-expected.txt


FAIL WebSockets: 'URL' assert_equals: expected (undefined) undefined but got (string) "ws://localhost:49001/"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:330:19
http://localhost:8800/websockets/interfaces/WebSocket/url/004.html:10:16
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/websockets/interfaces/WebSocket/url/004.html:8:5)
